Q2Q segment list simplification function
New function that allows neighbouring segments of the same cable type and thickness to be combined. This allows for more accurate data validation in the fault finding results.

Only used for the feedback and does not affect the number of line segments and the popup telling the user how far into the last segment the fault was located.