Main scene update
Changed the behaviour of the "RESTART CABLE SELECTION" button on the "Fault Location" screen to undo the last segment and allow the user to re-input a cable and values.

This is different to the standard "RESTART CABLE SELECTION" as this has to remove the last saved line segment to prevent extra segments being added to the list.