FinalResultPopupView update
Now uses the FaultPositionGuess class to accurately show feedback to the user. The users lines are presented via screenshot contrasting the expected lines from a perfect scenario (taken by the developers).

Uses validation to determine the colour and data in the strings given back to the user in the popup.

The user can use Next and Previous buttons to compare the different expected line segments. This works regardless of if the user inputs too many lines or wrong lines and relays colour coded feedback accordingly.